// reservoir/reservoir.types.generated/model508.ts
/**
 * Generated by orval v6.20.0 🍺
 * Do not edit manually.
 * beam-reservoir API
 * You are viewing the reference docs for the beam-reservoir API.        For a more complete overview with guides and examples, check out the <a href='https://reservoirprotocol.github.io'>Reservoir Protocol Docs</a>.
 * OpenAPI spec version: 5.296.1
 */
import type { Model507 } from './model507';
import type { Model508Kind } from './model508Kind';

export interface Model508 {
  action: string;
  description: string;
  /** Returns `nft-approval` or `order-signature` */
  id: string;
  items: Model507;
  /** Returns `request`, `signature`, or `transaction`. */
  kind: Model508Kind;
}